Fix IPv6 dual-stack handling
Previously, Forge had patches that would attempt to use IPv6 if it was provided as an IP from a hostname, however Java only returns IPv4 by default for legacy backwards-compatibility reasons and the system property being set was ignored as it's only checked once at startup.

This commit sets the appropriate system property at startup so that Java respects the OS' logic in determining what address to return first.

This allows for IPv6 to be used on dual-stack-resolving hosts which in turn can help improve ping times and reliability (the fastest endpoint is usually picked by the OS and if IPv4 is down but IPv6 is up, you can still connect with this fix, unlike before)